Refactor: DocumentListItem 핸들러 분리
